+++ CHANGELOG.md
@@ -22,6 +22,10 @@
   in ObjC mode for references to one-line declarations.  
   [John Fairhurst](https://github.com/johnfairh)
 
+* Fix crash with `` ` ` `` in markdown.  
+  [John Fairhurst](https://github.com/johnfairh)
+  [#1270](https://github.com/realm/jazzy/issues/1270)
+
 ## 0.13.7
 
 The next release of Jazzy will require a minimum of Ruby 2.6.